+++ src/android/MASPluginUser.java
@@ -11,6 +11,10 @@ import android.content.Context;
 import android.content.Intent;
 import android.util.Log;
 
+import com.ca.mas.foundation.MASAuthCredentialsAuthorizationCode;
+import com.ca.mas.foundation.MASAuthCredentialsJWT;
+import com.ca.mas.foundation.MASAuthCredentialsPassword;
+import com.ca.mas.foundation.MASAuthorizationResponse;
 import com.ca.mas.foundation.MASCallback;
 import com.ca.mas.foundation.MASFoundationStrings;
 import com.ca.mas.foundation.MASGroup;
@@ -31,10 +35,6 @@ import org.json.JSONObject;
 
 import static android.app.Activity.RESULT_CANCELED;
 import static android.app.Activity.RESULT_OK;
-
-/**
- * Created by trima09 on 12/13/2016.
- */
 
 public class MASPluginUser extends MASCordovaPlugin {
     private static final String TAG = MASPluginUser.class.getCanonicalName();
@@ -69,12 +69,22 @@ public class MASPluginUser extends MASCordovaPlugin {
                 loginWithUsernameAndPassword(args, callbackContext);
             } else if (action.equalsIgnoreCase("loginWithIdTokenAndTokenType")) {
                 loginWithIdTokenAndTokenType(args, callbackContext);
+            } else if (action.equalsIgnoreCase("loginWithAuthCode")) {
+                loginWithAuthCode(args, callbackContext);
+            } else if (action.equalsIgnoreCase("loginWithAuthCredentialsUsernamePassword")) {
+                loginWithAuthCredentialsUsernamePassword(args, callbackContext);
+            } else if (action.equalsIgnoreCase("loginWithAuthCredentialsJWT")) {
+                loginWithAuthCredentialsJWT(args, callbackContext);
+            } else if (action.equalsIgnoreCase("loginWithAuthCredentialsAuthCode")) {
+                loginWithAuthCredentialsAuthCode(args, callbackContext);
             } else if (action.equalsIgnoreCase("logoutUser")) {
                 logoutUser(callbackContext);
             } else if (action.equalsIgnoreCase("requestUserInfo")) {
                 requestUserInfo(callbackContext);
             } else if (action.equalsIgnoreCase("listAttributes")) {
                 listAttributes(callbackContext);
+            } else if (action.equalsIgnoreCase("getAuthCredentialsType")) {
+                getAuthCredentialsType(callbackContext);
             } else {
                 callbackContext.error("Invalid action");
                 return false;
@@ -128,12 +138,6 @@ public class MASPluginUser extends MASCordovaPlugin {
      * Fetches the current logged in MASUser and returns as json object
      */
     private void getCurrentUser(CallbackContext callbackContext) {
-        /*boolean retryOnNull = false;
-        try {
-            retryOnNull = args.getBoolean(0);
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }*/
         MASUser masUser = MASUser.getCurrentUser();
         if (masUser == null) {
             for (int i = 0; i < 60; i++) {
@@ -426,6 +430,124 @@ public class MASPluginUser extends MASCordovaPlugin {
     }
 
     /**
+     * Authenticates a user with an authorization code.
+     */
+    private void loginWithAuthCode(final JSONArray args, final CallbackContext callbackContext) {
+        String authorizationCode;
+        String state;
+        try {
+            authorizationCode = args.getString(0);
+            state = args.getString(1);
+        } catch (JSONException e) {
+            callbackContext.error(getError(e));
+            return;
+        }
+        MASAuthorizationResponse masAuthorizationResponse = new MASAuthorizationResponse(authorizationCode, state);
+
+        MASUser.login(masAuthorizationResponse, new MASCallback<MASUser>() {
+            @Override
+            public void onSuccess(MASUser masUser) {
+                String result = "Login with authorization code complete";
+                success(callbackContext, result, false);
+            }
+
+            @Override
+            public void onError(Throwable throwable) {
+                Log.e(TAG, throwable.getMessage(), throwable);
+                callbackContext.error(getError(throwable));
+            }
+        });
+    }
+
+    /**
+     * is used to login with provided username and password.
+     */
+    private void loginWithAuthCredentialsUsernamePassword(final JSONArray args, final CallbackContext callbackContext) {
+        String username;
+        String password;
+        try {
+            username = args.getString(0);
+            password = args.getString(1);
+        } catch (JSONException e) {
+            callbackContext.error(getError(e));
+            return;
+        }
+
+        MASAuthCredentialsPassword masAuthCredentialsPassword = new MASAuthCredentialsPassword(username, password.toCharArray());
+        MASUser.login(masAuthCredentialsPassword, new MASCallback<MASUser>() {
+            @Override
+            public void onSuccess(MASUser masUser) {
+                String result = "Login with authcredentials username and password complete";
+                success(callbackContext, result, false);
+            }
+
+            @Override
+            public void onError(Throwable throwable) {
+                Log.e(TAG, throwable.getMessage(), throwable);
+                callbackContext.error(getError(throwable));
+            }
+        });
+    }
+
+    /**
+     * is used to login with provided JWT and type.
+     */
+    private void loginWithAuthCredentialsJWT(final JSONArray args, final CallbackContext callbackContext) {
+        String jwt;
+        String type;
+        try {
+            jwt = args.getString(0);
+            type = args.getString(1);
+        } catch (JSONException e) {
+            callbackContext.error(getError(e));
+            return;
+        }
+        MASIdToken token = new MASIdToken.Builder().value(jwt).type(type).build();
+        MASAuthCredentialsJWT credentialsJWT = new MASAuthCredentialsJWT(token);
+        MASUser.login(credentialsJWT, new MASCallback<MASUser>() {
+            @Override
+            public void onSuccess(MASUser masUser) {
+                String result = "Login with JWT complete";
+                success(callbackContext, result, false);
+            }
+
+            @Override
+            public void onError(Throwable throwable) {
+                Log.e(TAG, throwable.getMessage(), throwable);
+                callbackContext.error(getError(throwable));
+            }
+        });
+    }
+
+
+    private void loginWithAuthCredentialsAuthCode(final JSONArray args, final CallbackContext callbackContext) {
+        String authCode = null;
+        String state = null;
+        try {
+            authCode = args.getString(0);
+            state = args.getString(1);
+        } catch (JSONException e) {
+            callbackContext.error(getError(e));
+            return;
+        }
+
+        MASAuthCredentialsAuthorizationCode credentials = new MASAuthCredentialsAuthorizationCode(authCode, state);
+        MASUser.login(credentials, new MASCallback<MASUser>() {
+            @Override
+            public void onSuccess(MASUser masUser) {
+                String result = "Login with Authorization Code using MASAuthCredentials complete";
+                success(callbackContext, result, false);
+            }
+
+            @Override
+            public void onError(Throwable throwable) {
+                Log.e(TAG, throwable.getMessage(), throwable);
+                callbackContext.error(getError(throwable));
+            }
+        });
+    }
+
+    /**
      * logs out the current logged in user
      */
     private void logoutUser(final CallbackContext callbackContext) {
@@ -489,6 +611,19 @@ public class MASPluginUser extends MASCordovaPlugin {
             Log.e(TAG, ex.getMessage(), ex);
             callbackContext.error(getError(ex));
         }
+    }
+
+    /**
+     * Returns the last authenticated session's type of auth credentials used.
+     */
+    private void getAuthCredentialsType(CallbackContext callbackContext) {
+        MASUser masUser = MASUser.getCurrentUser();
+        if (masUser == null) {
+            MASCordovaException e = new MASCordovaException(MASFoundationStrings.USER_NOT_CURRENTLY_AUTHENTICATED);
+            callbackContext.error(getError(e));
+            return;
+        }
+        success(callbackContext, MASUser.getAuthCredentialsType().toString(), false);
     }
 
     private JSONObject convertUserToJSModel(MASUser masUser) throws JSONException {