+++ www/MASPluginApplication.js
@@ -108,3 +108,4 @@ var MASPluginApplication = function() {
 }
 
 module.exports = MASPluginApplication;
+